requirements.txt
openai
async-openai
ipython
ipywidgets
asyncio 
aiofiles
python-dotenv
pydantic
